De-partial Legal class, move things to better spot
Item Restrictions now contains the "released" bool arrays.
Legal.cs is now a single class; all children features have been shuffled away to more appropriate classes.
