--- PKHeX.Core/Legality/Tables/Tables1.cs
@@ -1,9 +1,0 @@
-namespace PKHeX.Core;
-
-public static partial class Legal
-{
-    internal const int MaxSpeciesID_1 = 151;
-    internal const int MaxMoveID_1 = 165;
-    internal const int MaxItemID_1 = 255;
-    internal const int MaxAbilityID_1 = 0;
-}
